+++ src/Settings.Serializers.Json.Net/⬙/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ___
 
+## 3.3.0
+
+:calendar: _2023-03-28_
+
+### Fixed
+
+- The custom `EnumConverter` did not work for **Nullable** enumerations. Using such lead to an **InvalidOperationException**. This was due to the **System.Text.Json.JsonSerializer** not passing null values to the custom converter at all (see [here](https://learn.microsoft.com/en-us/dotnet/standard/serialization/system-text-json/converters-how-to?pivots=dotnet-6-0#handle-null-values)).
+
+### Removed
+
+- The already obsolete enumeration `IWriteOutOptions` **as comment** and **as property** have been removed.
+___
+
 ## 3.2.0
 
 :calendar: _2022-12-30_